requirements.txt
streamlit
augraphy
pillow
numpy